+++ src/js/CubicBezierCurve.js
@@ -20,7 +20,8 @@
  * @modified 2020-02-10 Fixed the translate(...) function (returning 'this' was missing).
  * @modified 2020-03-24 Ported this class from vanilla JS to Typescript.
  * @modified 2020-06-03 Added the getBounds() function.
- * @version 2.4.0
+ * @modified 2020-07-14 Changed the moveCurvePoint(...,Vertex) to moveCurvePoint(...,XYCoords), which is more generic.
+ * @version 2.4.1
  *
  * @file CubicBezierCurve
  * @public
@@ -69,7 +70,7 @@ var CubicBezierCurve = /** @class */ (function () {
      *
      * @method moveCurvePoint
      * @param {number} pointID - The numeric identicator of the point to move. Use one of the four eBezierPoint constants.
-     * @param {Vertex} moveAmount - The amount to move the specified point by.
+     * @param {XYCoords} moveAmount - The amount to move the specified point by.
      * @param {boolean} moveControlPoint - Move the control points along with their path point (if specified point is a path point).
      * @param {boolean} updateArcLengths - Specifiy if the internal arc segment buffer should be updated.
      * @instance