+++ src/ts/CubicBezierCurve.ts
@@ -19,7 +19,8 @@
  * @modified 2020-02-10 Fixed the translate(...) function (returning 'this' was missing).
  * @modified 2020-03-24 Ported this class from vanilla JS to Typescript.
  * @modified 2020-06-03 Added the getBounds() function.
- * @version 2.4.0
+ * @modified 2020-07-14 Changed the moveCurvePoint(...,Vertex) to moveCurvePoint(...,XYCoords), which is more generic.
+ * @version 2.4.1
  *
  * @file CubicBezierCurve
  * @public
@@ -152,7 +153,7 @@ export class CubicBezierCurve {
      *
      * @method moveCurvePoint
      * @param {number} pointID - The numeric identicator of the point to move. Use one of the four eBezierPoint constants.
-     * @param {Vertex} moveAmount - The amount to move the specified point by.
+     * @param {XYCoords} moveAmount - The amount to move the specified point by.
      * @param {boolean} moveControlPoint - Move the control points along with their path point (if specified point is a path point).
      * @param {boolean} updateArcLengths - Specifiy if the internal arc segment buffer should be updated.
      * @instance
@@ -160,7 +161,7 @@ export class CubicBezierCurve {
      * @return {void}
      **/
     moveCurvePoint( pointID:number,
-		    moveAmount:Vertex,
+		    moveAmount:XYCoords,
 		    moveControlPoint:boolean,
 		    updateArcLengths:boolean
 		  ) : void {