+++ src/ts/Vertex.ts
@@ -236,15 +236,6 @@ export class Vertex implements XYCoords, SVGSerializable {
      * @memberof Vertex
      **/
     add( x:number|XYCoords, y?:number ):Vertex {
-	/*
-	if( typeof x == 'number' ) {
-	    this.x += x;
-	    if( typeof y == 'number' )
-		this.y += y;
-	} else {
-	    this.x += x.x;
-	    this.y += x.y;
-	    }*/
 	if( typeof x == 'number' && typeof y == 'number' ) {
 	    this.x += x;
 	    this.y += y;